requirement.txt
streamlit
PyPDF2
python-docx
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
google-generativeai
